+++ app/src/main/java/site/madcat/movielibrary/ui/activity/MovieActivityContract.kt
@@ -1,9 +1,14 @@
 package site.madcat.movielibrary.ui.activity
 
+import androidx.lifecycle.MutableLiveData
 
 
 class MovieActivityContract {
+    enum class State { OK, ERROR }
+
     interface MovieActivityInterface {
+        val requestResult:MutableLiveData<String>
+
         fun onAttach(view: MovieActivity)
         fun onDetach()
         fun getMovie()
